Add an automatic commenter for new tooling PRs.
This workflow will comment on PRs that add new tools / implementations
to the website. It will do so only on PR open (i.e. once), not whenever
the PR is modified.

The goal is to direct implementers (or users submitting on their behalf)
to review some resources that might be applicable to them.

We can of course grow or shrink this list as needed, if we wish to point
to additional resources.

Right now, the way it detects which PRs are tool additions is
by considering any PR which touches:

    * the YAML files in the data directory
    * the pages/implementations/main.md page

to be possible candidates.

Additional tweaks could be made to add or remove these triggers if the
commenter is too noisy, or not noisy enough.