Add a proposed mechanism for tracking known "quality" implementations.
There is obvious overlap here with our existing list of implementations
-- note though that we have loose (or no) criteria for listing
implementations on the page.

Here instead we wish to specifically raise languages for which we know
no good implementation to exist, despite the language being popular or
widespread, if any.

What's here does *not* yet render this data anywhere, nor does it
implement the dynamic API calling mentioned in the body (for retrieving
statistics from any listed repository).

What it does do is identify a simple initial list of languages (from
TIOBE and GitHub's own lists), and propose a format for tracking data
points about implementations meeting the proposed criteria.

The initial goal is to get feedback on this template, and to solicit
help in filling it in for these languages.

Refs: json-schema-org/community#418