chore(package.json): Update dependencies and node version (#727)
* patch!: use the version of node in the documentation

So, hear me out. Inside README.md you claim that the project needs node v20.9.0+.
But package.json specifies that it needs to be EXACTLY node 20.0.0

I updated the field in the package.json to use the correct version of
node. Or range of versions, for that matter.

* chore(yarn): use latest stable yarn

Yarn 4 is so much faster and more efficient than yarn 1, and its
dependency resolution algorythm is more secure.

But I understand if you would like to keep the older versions.

* patch(ci): did you know `actions/setup-node` can do caching for you?

* fix: add `moment-timezone` as a dependency

* ci: update caching accordingly for next builds

* merge: resolved conflicts

* revert: Restored the `yarn install`

* patch(package.json): pin node to `^20.9.0`

* patch: the second yarn install

* fix: pinned node version to 20.9.0 in all github actions

* fix: ah yes I forgot about the `@types`

thanks for that one

Co-authored-by: Alok Gupta <alokgupta1560@gmail.com>

* fix: removed the next-plausible and updated lockfile

* patch: updated lockfile after upstream update

---------

Co-authored-by: Alok Gupta <alokgupta1560@gmail.com>